Change modules to contain package reference
This change ensures that all Module objects passed around have a
valid Package reference even when they are not part of a package.
The goal is to standardize on having the Package attached to
simplify the branching logic that would have been required to
implement certain features. Ex: if module.package is not None...

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>